fix: error "cannot be named without a reference" on both files:
- auth.ts > line 18 (signIn) error:
```
The inferred type of 'signIn' cannot be named without a reference to 'next-auth/node_modules/@auth/core/providers'. This is likely not portable. A type annotation is necessary.
```
  - error connected to social.tsx:
    - social.tsx > onlick function changed - improved for this error removal.

- auth.config.ts > "export default" error:
```
The inferred type of 'default' cannot be named without a reference to 'next-auth/node_modules/@auth/core/providers'. This is likely not portable. A type annotation is necessary.
```
  - export default changed to "const config: NextAuthConfig = {" and "} as const; export default config;" added to end of file.